refactor: 컴포넌트를 각각 서브시스템 아래로 이동
