+++ src/pages/map/component/FetchTypeToggle.js
