+++ src/pages/search/component/BuildingSearchResult.js
